+++ CHANGELOG.markdown
@@ -2,6 +2,7 @@
 
 ## Unreleased
 
+- Add alphabetical ordering for private and protected methods to ClassOrganization.
 - Fix ClassOrganization adding blank lines between consecutive macros.
 - Fix ClassOrganization autocorrect dropping uncategorizable nodes.
 - Fix DescribeBlockOrder sorting controller action names outside of controllers.